First version of a molecular orbital user interface
Still quite a few rough edges, but you can select an orbital, the cube
will be requested, and shown once ready. It selects the HOMO but doesn't
show it right now, and the isosurface value isn't exposed. This can also
take a while, and offers no visual feedback which isn't ideal.