Show SVG for calculation if no 3D coordinates
Just like the case for molecules, in the calculation pages,
show an SVG instead of a 3D view if there are no 3D coordinates
in the molecule.

This seems to work, although I think I noticed an unrelated bug
that chemml calculations do not appear to show 3D structure,
even when a 3D structure is present.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>