Allow for advanced searches for molecules
In an advanced search, you can do things like the following:

`mass >= 200 and atomCount <= 50`

The way I currently have it set up, spaces are not required except
for around the logical operators 'and' and 'or' (to distinguish them
from any substrings in a string). '==' is the equality operator,
because you can also do things like:

`smiles == C1=CC=CC=C1`

And we need to be able to distinguish between '==' and '=' in smiles.

Let me know if I should change any of these things.

The fields that are searchable are: inchi, inchikey, name, formula,
smiles, mass, atomCount, and heavyAtomCount.

This depends on [this PR](https://github.com/OpenChemistry/mongochemserver/pull/140).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>