A little bit of UI and code cleanup
Starting using a TextField instead of Input for the image name,
since that appears nicer and allows for a label.

Added theme-based margin spacing for the components.

Call this.props.onPull() directly instead of setting it and
calling it.

Added a line break after the image name, so that the container
type and image name appear on the same line, and "Pull Image"
appears on a new line.

Removed an unneeded import.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>