Added basic and advanced tabs for pulling images
The basic tab just pulls a single image. The advanced tab will pull
a list of images.

The default images in the advanced tab are a default list of images
that are needed to run the examples.

This works in my testing. However, one cumulus task is launched for
each image, which is not preferred. It would be better if a single
cumulus task was launched for all of the images simultaneously.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>